+++ src/commands/add.rs
@@ -2,7 +2,8 @@ use crate::commands::helpers::database::{self, MessageType};
 use crate::commands::helpers::time::{self, MinusOffsetChoice, PlusOffsetChoice};
 use crate::commands::helpers::tracking;
 use crate::config::{BloomBotEmbed, CHANNELS, EMOJI};
-use crate::database::{DatabaseHandler, TrackingProfile};
+use crate::data::tracking_profile::TrackingProfile;
+use crate::database::DatabaseHandler;
 use crate::events;
 use crate::Context;
 use anyhow::{Context as AnyhowContext, Result};
