+++ src/commands/helpers/mod.rs
@@ -1,5 +1,5 @@
 pub(super) mod courses;
 pub(super) mod database;
 pub mod pagination;
-pub(super) mod time;
+pub mod time;
 pub(super) mod tracking;
