+++ src/events/helpers/leaderboards.rs
@@ -2,8 +2,10 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use crate::charts;
+use crate::commands::helpers::time::Timeframe;
 use crate::commands::stats::{LeaderboardType, SortBy};
-use crate::database::{DatabaseHandler, LeaderboardUserStats, Timeframe};
+use crate::data::stats::LeaderboardUserStats;
+use crate::database::DatabaseHandler;
 use anyhow::Result;
 use log::{error, info};
 use poise::serenity_prelude::{GuildId, Http, UserId};
